requirements.txt
requests>=2.31.0
feedparser>=6.0.10
newspaper4k>=0.9.3
beautifulsoup4>=4.12.0
openai>=1.12.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pyyaml>=6.0.1
sparkpost>=1.3.10
python-dateutil>=2.8.2
pytest>=8.0.0
